Updated styling
fixed feedback on the concessions tab such as moving the image to the bottom, centering the add to order button, adding the number order to the image and added functionality to where they can add how many they want to order. Also added where they can remove the item. Also fixed to where they log in and see movies first. Also fixed the the tab icon not working and fixed the mapping issues on locations.