Updated functionality
Updated the tickets tab by removing the qr scanner at the top and added a qr code for them to scan when they arrive at the movie. For the theaters tab updated the locations based on the client given locations and also added functionality where it gets the neartest theater that is located to them.